修改实体枚举类型

+++ HumanResourcesManagementBackend.Repository/Migrations/Configuration.cs
@@ -32,7 +32,7 @@
                     Password = "admin".Encrypt(),
                     Question = string.Empty,
                     Answer = string.Empty,
-                    Status = (int)DataStatus.Enable
+                    Status = DataStatus.Enable
                 });
             }
         }
